@@ -1,6 +1,8 @@
 import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
 import {AuthenticationService} from '../_services/authentication.service';
+import {decode} from 'punycode';
+import {Role} from '../_models/role';
 
 @Injectable({
   providedIn: 'root'
@@ -15,10 +17,14 @@ export class AuthGuard implements CanActivate {
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     const currentUser = this.authenticationService.currentUserValue;
-    const userRoles = this.authenticationService.authorities;
 
     if (currentUser) {
       // check if route is restricted by role
+      const userRoles = currentUser.authorities;
+      if (userRoles.includes('ROLE_ADMIN')) {
+        return true;
+      }
+
       if (route.data.roles) {
         for (const id in route.data.roles) {
           if (userRoles.includes(route.data.roles[id])) {
