@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  api: 'http://localhost:8081/api',
-  auth: 'http://localhost:8081/auth',
+  api: 'https://jaronsky.ddns.net:8085/api',
+  auth: 'https://jaronsky.ddns.net:8085/auth',
 };
