@@ -4,8 +4,8 @@
 
 export const environment = {
   production: false,
-  api: 'http://localhost:8081/api',
-  auth: 'http://localhost:8081/auth',
+  api: 'http://localhost:8086/api',
+  auth: 'http://localhost:8086/auth',
 };
 
 /*
